Return null as default client id instead of "anon"
Nulls are handled by RateLimitMiddleware.ResolveIdentity when iterating
over client resolvers to find a match, while "anon" was treated as a
valid id and thus causing the iteration to stop and custom resolvers to
never be called.

This should not modify behavior as "anon" is still returned as default
client id if no resolvers have a match.